% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.fun.R
\name{add.fun}
\alias{add.fun}
\title{This function adds two numbers together.}
\usage{
add.fun(x, y)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
This function adds two numbers together.
}
