% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.fun.R
\name{add.fun}
\alias{add.fun}
\title{Add together two numbers.}
\usage{
add.fun(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
\examples{
add.fun(2, 3)
add.fun(5, 5)

}
